+++ BackEnd/TgLab.Domain/Models/Bet.cs
@@ -6,9 +6,9 @@ namespace TgLab.Domain.Models
     {
         public int Id { get; set; }
         public int WalletId { get; set; }
-        public int Amount { get; set; }
+        public decimal Amount { get; set; }
         public BetStage Stage { get; set; }
-        public int Bounty {  get; set; } 
+        public decimal Bounty {  get; set; } 
         public DateTime Time { get; set; }
 
         public Wallet Wallet { get; set; }
